backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pdfplumber>=0.10.0
google-cloud-vision>=3.7.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0
email-validator>=2.1.0
